feat: embed client dir in builds
Replace the locally-linked client dir with an embedded one to make
builds a little more portable for my use. This retains support
for pointing smocker at a locally-provided client dir but uses the
embedded one by default if it contains an index.html. Update builds
as well to ensure that yarn runs first in a release build, providing
the embedded files needed for the Go build. The smocker.tar.gz
product no longer includes the client dir since it's not required
and the likelihood of someone wanting to modify mostly-generated
web files is unlikely.

gitignore rules for the build dir have been modified to keep a
single .gitignore file in build/client, ensuring the directory
exists during `go build` and allowing it to be embedded, even if
primarily empty. The .gitignore file is not embedded.